Refactored the login page to be in the login class, and moved the authentication to the backend
