+++ .gitignore
@@ -20,3 +20,4 @@
 /src/movementHandler.o
 /src/TextEditor.o
 /src/abc
+/src/authentication.o
