+++ src/graphic/graphic.cpp
@@ -38,18 +38,6 @@ using std::ofstream;
 #define IMGUI_DEFINE_MATH_OPERATORS
 #include "imgui_internal.h"
 
-static void HelpMarker(const char* desc)
-{
-    ImGui::TextDisabled(ICON_FA_CIRCLE_QUESTION);
-    if (ImGui::IsItemHovered(ImGuiHoveredFlags_DelayShort))
-    {
-        ImGui::BeginTooltip();
-        ImGui::PushTextWrapPos(ImGui::GetFontSize() * 35.0f);
-        ImGui::TextUnformatted(desc);
-        ImGui::PopTextWrapPos();
-        ImGui::EndTooltip();
-    }
-}
 
 void graphic::setup(){
         TextEditor editor;
@@ -185,10 +173,12 @@ void graphic::setup(){
     ImGui_ImplOpenGL3_Init(glsl_version);
 
     imageHandler image = imageHandler();
-    database& db = database::getInstance();
-    //db.connect();
     characterManager character = characterManager();
     characterBuilder builder = characterBuilder(&image);
+
+    database& db = database::getInstance();
+    //db.connect();
+    login Login = login(width_px, height_px);
 
     string pathMap;
 #if defined(__APPLE__)
@@ -241,7 +231,7 @@ void graphic::setup(){
 
         if(show_login)
         {
-            makeLogIn();
+            makeLogIn(Login);
         }
 
         if(show_display){
@@ -530,274 +520,23 @@ string graphic::executeCPP(string code){
     return executeOutput;
 }
 
-void graphic::makeLogIn()
+void graphic::makeLogIn(login& Login)
 {
     ImGuiStyle& style = ImGui::GetStyle();
     style.FrameRounding = 7.5f;
     style.Colors[ImGuiCol_Text] = ImVec4(0.0f, 0.0f, 0.0f, 1.00f);
 
-    ImGui::SetNextWindowSize({(float)width_px, (float)height_px});
-    ImGui::SetNextWindowPos({0, 0});
-
-    // coordinates for the top left corner
-    static float minWidth = (width_px / 2.f) - 200.f;
-    static float minHeight = (height_px / 2.f) - 240.f;
-    static float txtPos_x = 0.f;
-
-    static login Login = login();
-
-    static char username[64] = "";
-    static char passwd[64] = "";
-
-    static char createUsername[64] = "";
-    static char createPasswd[64] = "";
-    static char confirmPasswd[64] = "";
-
-    static bool createAccount = false;
-    static bool viewPasswd = false; // keep password hidden by default
-    static bool errorAuth = false;
-    static bool errorCmp = false;
-    static bool errorCreate = false;
-
-    ImGui::Begin("Log In", NULL, ImGuiWindowFlags_NoMove | ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoCollapse | ImGuiWindowFlags_NoTitleBar | ImGuiWindowFlags_NoBackground);
+    Login.drawLoginScreen();
+    if(Login.checkAuth())
     {
-        ImDrawList* draw_list = ImGui::GetWindowDrawList();
-        draw_list->AddRectFilled(ImVec2(minWidth, minHeight), ImVec2(minWidth + 400.f, minHeight + 500.f), ImColor(ImVec4(1.0f, 1.0f, 1.0f, 1.0f)), 20.0f);
+        show_display = true;
+        show_process = true;
+        show_config = true;
+        show_charSelector = true;
 
-        if(!createAccount)
-        {
-            txtPos_x = (width_px / 2.f) - ((ImGui::CalcTextSize("LOGIN").x) / 2.f);
-            ImGui::SetCursorPos(ImVec2(txtPos_x, minHeight + 30.f));
-            ImGui::Text("LOGIN");
+        style.FrameRounding = 0.f;
+        style.Colors[ImGuiCol_Text] = ImVec4(1.0f, 1.0f, 1.0f, 1.00f);
 
-            txtPos_x = (width_px / 2.f) - ((ImGui::CalcTextSize("Username").x) / 2.f);
-            ImGui::SetCursorPos(ImVec2(txtPos_x, minHeight + 120.f));
-            ImGui::Text("Username");
-
-            ImGui::SetCursorPos(ImVec2(minWidth + 50.f, minHeight + 140.f));
-            ImGui::PushItemWidth(300);
-            ImGui::InputText(" ",username, IM_ARRAYSIZE(username), ImGuiInputTextFlags_None);
-            ImGui::PopItemWidth();
-
-            txtPos_x = (width_px / 2.f) - ((ImGui::CalcTextSize("Password").x) / 2.f);
-            ImGui::SetCursorPos(ImVec2(txtPos_x, minHeight + 180.f));
-            ImGui::Text("Password");
-
-            ImGui::SetCursorPos(ImVec2(minWidth + 50.f, minHeight + 200.f));
-            ImGui::PushItemWidth(300);
-            if(!viewPasswd)
-            {
-                ImGui::InputText("  ",passwd, IM_ARRAYSIZE(passwd), ImGuiInputTextFlags_Password);
-                ImGui::PopItemWidth();
-
-                ImGui::SetCursorPos(ImVec2(minWidth + 350.f, minHeight + 201.f));
-                ImGui::PushID(8);
-                ImGui::PushStyleColor(ImGuiCol_Button, (ImVec4)ImColor::HSV(0.f / 360.f,0.0f,1.0f));
-                if(ImGui::Button(ICON_FA_EYE_SLASH, ImVec2(30.f,16.f)))
-                {
-                    viewPasswd = true;
-                }
-                ImGui::PopStyleColor(1);
-                ImGui::PopID();
-            }
-            else
-            {
-                ImGui::InputText("  ",passwd, IM_ARRAYSIZE(passwd), ImGuiInputTextFlags_None);
-                ImGui::PopItemWidth();
-
-                ImGui::SetCursorPos(ImVec2(minWidth + 350.f, minHeight + 201.f));
-                ImGui::PushID(9);
-                ImGui::PushStyleColor(ImGuiCol_Button, (ImVec4)ImColor::HSV(0.f / 360.f,0.0f,1.0f));
-                if(ImGui::Button(ICON_FA_EYE, ImVec2(30.f,16.f)))
-                {
-                    viewPasswd = false;
-                }
-                ImGui::PopStyleColor(1);
-                ImGui::PopID();
-            }
-
-            txtPos_x = (width_px / 2.f) - ((ImGui::CalcTextSize("Don't have an account?").x) / 2.f);
-            ImGui::SetCursorPos(ImVec2(txtPos_x - 50.f, minHeight + 320.f));
-            ImGui::Text("Don't have an account?");
-            ImGui::SameLine();
-            ImGui::PushID(2);
-            ImGui::PushStyleColor(ImGuiCol_Button, (ImVec4)ImColor::HSV(0.f / 360.f,0.0f,0.71f));
-            ImGui::PushStyleColor(ImGuiCol_ButtonHovered, (ImVec4)ImColor::HSV(0.f / 360.f,0.0f,0.33f));
-            ImGui::PushStyleColor(ImGuiCol_ButtonActive, (ImVec4)ImColor::HSV(200.f / 360.f,1.0f,0.61f));
-            if(ImGui::Button("Sign Up", ImVec2(100.f,20.f)))
-            {
-                createAccount = true;
-                viewPasswd = false;
-                errorAuth = false;
-            }
-            ImGui::PopStyleColor(3);
-            ImGui::PopID();
-
-            ImGui::SetCursorPos(ImVec2((width_px / 2.f - 50.f),minHeight + 380.f));
-            ImGui::PushID(1);
-            ImGui::PushStyleColor(ImGuiCol_Button, (ImVec4)ImColor::HSV(200.f / 360.f,1.0f,1.0f));
-            ImGui::PushStyleColor(ImGuiCol_ButtonHovered, (ImVec4)ImColor::HSV(200.f / 360.f,0.8f,1.0f));
-            ImGui::PushStyleColor(ImGuiCol_ButtonActive, (ImVec4)ImColor::HSV(200.f / 360.f,1.0f,0.81f));
-            if(ImGui::Button("Login", ImVec2(100.f,50.f)))
-            {
-                if(Login.inputValidation( username, passwd, true)) // this will validate the input and then authenticate the user
-                {
-                    show_display = true;
-                    show_process = true;
-                    show_config = true;
-                    show_charSelector = true;
-
-                    style.FrameRounding = 0.f;
-                    style.Colors[ImGuiCol_Text] = ImVec4(1.0f, 1.0f, 1.0f, 1.00f);
-
-                    show_login = false;
-                }
-                else
-                {
-                    errorAuth = true;
-                }
-            }
-
-            if(errorAuth)
-            {
-                draw_list->AddRectFilled(ImVec2(minWidth + 30.f, minHeight + 235.f), ImVec2(minWidth + 370.f, minHeight + 290.f), ImColor(ImVec4(1.0f, 0.0f, 0.0f, 1.0f)), 20.0f);
-                ImGui::PushStyleColor(ImGuiCol_Text, IM_COL32(0,0,0,255));
-                ImGui::SetCursorPos(ImVec2(minWidth + 40.f, minHeight + 244.f));
-                ImGui::Text(ICON_FA_TRIANGLE_EXCLAMATION);
-                txtPos_x = (width_px / 2.f) - ((ImGui::CalcTextSize("Oops! Something went wrong.").x) / 2.f);
-                ImGui::SetCursorPos(ImVec2(txtPos_x, minHeight + 242.f));
-                ImGui::Text("Oops! Something went wrong.");
-                txtPos_x = (width_px / 2.f) - ((ImGui::CalcTextSize("Please double-check your username and password.").x) / 2.f);
-                ImGui::SetCursorPos(ImVec2(txtPos_x, minHeight + 265.f));
-                ImGui::Text("Please double-check your username and password.");
-                ImGui::PopStyleColor();
-            }
-            ImGui::PopStyleColor(3);
-            ImGui::PopID();
-        }
-        else
-        {
-            ImGui::SetCursorPos(ImVec2(minWidth + 20.f, minHeight + 20.f));
-            ImGui::PushID(7);
-            ImGui::PushStyleColor(ImGuiCol_Button, (ImVec4)ImColor::HSV(0.f / 360.f,0.0f,1.0f));
-            if(ImGui::Button(ICON_FA_CHEVRON_LEFT, ImVec2(23.f,18.f)))
-            {
-                createAccount = false;
-                viewPasswd = false; // turn off both the new account and viewing passwd
-            }
-            ImGui::PopStyleColor(1);
-            ImGui::PopID();
-
-            txtPos_x = (width_px / 2.f) - ((ImGui::CalcTextSize("Create Your Hash Harbor Account").x) / 2.f);
-            ImGui::SetCursorPos(ImVec2(txtPos_x, minHeight + 30.f));
-            ImGui::Text("Create Your Hash Harbor Account");
-
-            txtPos_x = (width_px / 2.f) - ((ImGui::CalcTextSize("Username").x) / 2.f);
-            ImGui::SetCursorPos(ImVec2(txtPos_x, minHeight + 120.f));
-            ImGui::Text("Username");
-            ImGui::PushStyleColor(ImGuiCol_Text, IM_COL32(255,255,255,255));
-            ImGui::SameLine(); HelpMarker("A username should be between 3 and 40 character\nA Username may contain:\n  Lowercase letters: a-z\n  Uppercase letters: A-Z\n  Digits: 0-9\n  Special character: - _ .");
-            ImGui::PopStyleColor();
-
-            ImGui::SetCursorPos(ImVec2(minWidth + 50.f, minHeight + 140.f));
-            ImGui::PushItemWidth(300);
-            ImGui::InputText(" ",createUsername, IM_ARRAYSIZE(createUsername), ImGuiInputTextFlags_None);
-            ImGui::PopItemWidth();
-
-            txtPos_x = (width_px / 2.f) - ((ImGui::CalcTextSize("Password").x) / 2.f);
-            ImGui::SetCursorPos(ImVec2(txtPos_x, minHeight + 180.f));
-            ImGui::Text("Password");
-            ImGui::PushStyleColor(ImGuiCol_Text, IM_COL32(255,255,255,255));
-            ImGui::SameLine(); HelpMarker("Password must be at least 8 characters\nPassword must be no more than 40 character\nPassword MUST contain at least one from each group below:\n  lowercase letter: a-z\n  Uppercase letter: A-Z\n  Digit: 0-9\n  Special character: ~ ` ! @ # $ % ^ & * ( ) _ - + = \n                   { } : ; < > ?");
-            ImGui::PopStyleColor();
-
-            ImGui::SetCursorPos(ImVec2(minWidth + 50.f, minHeight + 200.f));
-            ImGui::PushItemWidth(300);
-            if(!viewPasswd)
-            {
-                ImGui::InputText("  ",createPasswd, IM_ARRAYSIZE(createPasswd), ImGuiInputTextFlags_Password);
-                ImGui::PopItemWidth();
-
-                ImGui::SetCursorPos(ImVec2(minWidth + 350.f, minHeight + 201.f));
-                ImGui::PushID(5);
-                ImGui::PushStyleColor(ImGuiCol_Button, (ImVec4)ImColor::HSV(0.f / 360.f,0.0f,1.0f));
-                if(ImGui::Button(ICON_FA_EYE_SLASH, ImVec2(30.f,16.f)))
-                {
-                    viewPasswd = true;
-                }
-                ImGui::PopStyleColor(1);
-                ImGui::PopID();
-
-                txtPos_x = (width_px / 2.f) - ((ImGui::CalcTextSize("Confirm Password").x) / 2.f);
-                ImGui::SetCursorPos(ImVec2(txtPos_x, minHeight + 240.f));
-                ImGui::Text("Confirm Password");
-
-                ImGui::SetCursorPos(ImVec2(minWidth + 50.f, minHeight + 260.f));
-                ImGui::PushItemWidth(300);
-                ImGui::InputText("   ",confirmPasswd, IM_ARRAYSIZE(confirmPasswd), ImGuiInputTextFlags_Password);
-                ImGui::PopItemWidth();
-            }
-            else
-            {
-                ImGui::InputText("  ",createPasswd, IM_ARRAYSIZE(createPasswd), ImGuiInputTextFlags_None);
-                ImGui::PopItemWidth();
-
-                ImGui::SetCursorPos(ImVec2(minWidth + 350.f, minHeight + 201.f));
-                ImGui::PushID(6);
-                ImGui::PushStyleColor(ImGuiCol_Button, (ImVec4)ImColor::HSV(0.f / 360.f,0.0f,1.0f));
-                if(ImGui::Button(ICON_FA_EYE, ImVec2(30.f,16.f)))
-                {
-                    viewPasswd = false;
-                }
-                ImGui::PopStyleColor(1);
-                ImGui::PopID();
-
-                txtPos_x = (width_px / 2.f) - ((ImGui::CalcTextSize("Confirm Password").x) / 2.f);
-                ImGui::SetCursorPos(ImVec2(txtPos_x, minHeight + 240.f));
-                ImGui::Text("Confirm Password");
-
-                ImGui::SetCursorPos(ImVec2(minWidth + 50.f, minHeight + 260.f));
-                ImGui::PushItemWidth(300);
-                ImGui::InputText("   ",confirmPasswd, IM_ARRAYSIZE(confirmPasswd), ImGuiInputTextFlags_None);
-                ImGui::PopItemWidth();
-            }
-
-            ImGui::SetCursorPos(ImVec2((width_px / 2.f - 50.f),minHeight + 380.f));
-            ImGui::PushID(3);
-            ImGui::PushStyleColor(ImGuiCol_Button, (ImVec4)ImColor::HSV(200.f / 360.f,1.0f,1.0f));
-            ImGui::PushStyleColor(ImGuiCol_ButtonHovered, (ImVec4)ImColor::HSV(200.f / 360.f,0.44f,1.0f));
-            ImGui::PushStyleColor(ImGuiCol_ButtonActive, (ImVec4)ImColor::HSV(200.f / 360.f,1.0f,0.61f));
-            if(ImGui::Button("Sign Up", ImVec2(100.f,50.f)))
-            {
-                errorCmp = false;
-                errorCreate = false;
-                if(strcmp(createPasswd,confirmPasswd) != 0) // confirm passwords match
-                {
-                    errorCmp = true;
-                    createPasswd[0] = '\0';
-                    confirmPasswd[0] = '\0';
-                }
-                else if(!Login.inputValidation( createUsername, createPasswd, false)) // if account passes input validation, hashing and adding to database
-                {
-                    errorCreate = true;
-                }
-                else
-                {
-                    show_display = true;
-                    show_process = true;
-                    show_config = true;
-                    show_charSelector = true;
-
-                    style.FrameRounding = 0.f;
-                    style.Colors[ImGuiCol_Text] = ImVec4(1.0f, 1.0f, 1.0f, 1.00f);
-
-                    show_login = false;
-                }
-            }
-            ImGui::PopStyleColor(3);
-            ImGui::PopID();
-        }
-
+        show_login = false;
     }
-    ImGui::End();
 }
